fixed wrong kompressor key fixed help command showing too much removed admin commands as they are not usefull anymore
