+++ src/main/java/de/blackcraze/grb/util/StandingDataInitializer.java
@@ -48,6 +48,20 @@ public class StandingDataInitializer {
 				getStockTypeDao().save(type);
 				System.out.println("Created new stock type: " + type.getName());
 			}
+			
+			for (StockType type : stocks) {
+				boolean found = false;
+				for (CSVRecord record : records) {
+					String name = record.get(0);
+					if(type.getName().equals(name)){
+						found = true;
+					}
+				}
+				if(!found){
+					System.out.println("Delete obsoleth stock type: " + type.getName());
+					getStockTypeDao().delete(type);
+				}
+			}
 		} catch (Throwable e) {
 			e.printStackTrace();
 		} finally {
