+++ src/main/resources/StockTypes.csv
@@ -81,7 +81,7 @@ GEAR,18500
 OIL,21
 POLISHED_OBSIDIAN,280
 OBSIDIAN_KNIFE,12500
-KOMPRESSOR,44000
+COMPRESSOR,44000
 HELIUM_3,20
 OBSIDIAN,20
 LIQUID_NITROGEN,12500
